user metadata refactor
