+++ console/src/app/pages/users/user-detail/metadata/metadata.component.ts
@@ -1,55 +1,16 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { Metadata } from 'src/app/proto/generated/zitadel/metadata_pb';
-import { ManagementService } from 'src/app/services/mgmt.service';
-import { ToastService } from 'src/app/services/toast.service';
-
-import { MetadataDialogComponent } from '../metadata-dialog/metadata-dialog.component';
 
 @Component({
   selector: 'cnsl-metadata',
   templateUrl: './metadata.component.html',
   styleUrls: ['./metadata.component.scss'],
 })
-export class MetadataComponent implements OnInit {
-  @Input() userId: string = '';
-  public metadata: Metadata.AsObject[] = [];
-  public loading: boolean = false;
+export class MetadataComponent {
+  @Input() public metadata: Metadata.AsObject[] = [];
+  @Input() public disabled: boolean = false;
+  @Input() public loading: boolean = false;
+  @Output() public editClicked: EventEmitter<void> = new EventEmitter();
 
-  constructor(private dialog: MatDialog, private service: ManagementService, private toast: ToastService) {}
-
-  ngOnInit(): void {
-    this.loadMetadata();
-  }
-
-  public editMetadata(): void {
-    const dialogRef = this.dialog.open(MetadataDialogComponent, {
-      data: {
-        userId: this.userId,
-      },
-    });
-
-    dialogRef.afterClosed().subscribe(() => {
-      this.loadMetadata();
-    });
-  }
-
-  public loadMetadata(): Promise<any> {
-    this.loading = true;
-    return (this.service as ManagementService)
-      .listUserMetadata(this.userId)
-      .then((resp) => {
-        this.loading = false;
-        this.metadata = resp.resultList.map((md) => {
-          return {
-            key: md.key,
-            value: atob(md.value as string),
-          };
-        });
-      })
-      .catch((error) => {
-        this.loading = false;
-        this.toast.showError(error);
-      });
-  }
+  constructor() {}
 }
