+++ console/src/app/pages/users/user-detail/user-detail/user-detail.component.html
@@ -206,7 +206,12 @@
         </ng-container>
 
         <ng-container *ngIf="currentSetting && currentSetting === 'metadata'">
-          <cnsl-metadata *ngIf="user" [userId]="user.id"></cnsl-metadata>
+          <cnsl-metadata
+            [metadata]="metadata"
+            [disabled]="(['user.write:' + user.id, 'user.write'] | hasRole | async) === false"
+            *ngIf="user && user.id"
+            (editClicked)="editMetadata()"
+          ></cnsl-metadata>
         </ng-container>
       </div>
     </cnsl-sidenav>
